created quanticSearchBoxInput with customEvents, cleaned up standalone and default searchbox
