--- packages/quantic/force-app/main/default/lwc/quanticSearchBox/templates/expandableSearchBox.css
@@ -1,38 +1,0 @@
-@import 'c/searchBoxStyle';
-
-.searchbox__container {
-  overflow: visible;
-  border: var(--lwc-borderWidthThin, 1px) solid var(--lwc-brandAccessible, #0176d3);
-}
-
-.z-index-high {
-  z-index: 20;
-}
-
-textarea.searchbox__input {
-  padding-top: 0.8rem;
-  padding-bottom: 0.8rem;
-  line-height: 20px;
-  max-height: 8.5rem;
-  resize: none;
-  border: none;
-  box-shadow: none;
-  overflow-x: clip;
-  margin-right: 0.8rem;
-}
-
-.searchbox__submit-button {
-  border: none;
-  border-radius: 0 var(--lwc-borderRadiusMedium, 0.25rem) var(--lwc-borderRadiusMedium, 0.25rem) 0;
-}
-
-.searchbox__container-wrapper {
-  min-height: 3.125rem;
-}
-
-.searchbox_floating-container {
-  position: absolute;
-  width: 100%;
-  background-color: var(--lwc-colorBackgroundAlt, rgb(255, 255, 255));
-  border-radius: 0.25rem;
-}
