+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/templates/searchBox.css
