+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/templates/searchBox.html
@@ -1,0 +1,9 @@
+<template>
+  <c-quantic-search-box-input
+    textarea={textarea}
+    input-value={inputValue}
+    without-submit-button={withoutSubmitButton}
+    placeholder={placeholder}
+    suggestions={suggestions}
+  ></c-quantic-search-box-input>
+</template>
