+++ packages/quantic/force-app/main/default/lwc/quanticSearchBoxInput/__tests__/quanticSearchBoxInput.test.js
@@ -1,0 +1,10 @@
+import { createElement } from 'lwc';
+import QuanticSearchBoxInput from 'c/quanticSearchBoxInput';
+
+describe('c-quantic-search-box-input', () => {
+// 2 use cases : default and expandable
+
+    it('something', () => {
+
+    });
+});
