+++ packages/quantic/force-app/main/default/lwc/quanticSearchBoxInput/quanticSearchBoxInput.js-meta.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
+    <apiVersion>58.0</apiVersion>
+    <isExposed>false</isExposed>
+</LightningComponentBundle>
