+++ packages/quantic/force-app/main/default/lwc/quanticSearchBoxInput/templates/defaultSearchBoxInput.css
@@ -1,0 +1,9 @@
+@import 'c/searchBoxStyle';
+
+.searchbox__input-container {
+  display: flex;
+}
+
+input {
+  border: none;
+}
