+++ packages/quantic/force-app/main/default/lwc/quanticSearchBoxInput/templates/expandableSearchBoxInput.css
@@ -1,5 +1,9 @@
 @import 'c/searchBoxStyle';
 
+.searchbox__clear-button-container {
+  margin: inherit;
+}
+
 .searchbox__container {
   overflow: visible;
   border: var(--lwc-borderWidthThin, 1px) solid var(--lwc-brandAccessible, #0176d3);
@@ -18,6 +22,8 @@ textarea.searchbox__input {
   border: none;
   box-shadow: none;
   overflow-x: clip;
+  /* there for expandable, not default */
+  margin-right: 0.8rem;
 }
 
 .searchbox__submit-button {
