+++ packages/quantic/force-app/main/default/lwc/quanticSearchBoxInput/templates/expandableSearchBoxInput.html
@@ -1,4 +1,3 @@
-<!-- sldsValidatorIgnore -->
 <template>
   <div class="slds-form-element z-index-high searchbox__container-wrapper">
     <div class="slds-form-element__control searchbox_floating-container">
@@ -7,14 +6,16 @@
           class="slds-combobox slds-dropdown-trigger slds-dropdown-trigger_click"
           aria-expanded="false"
           aria-haspopup="listbox"
-          role="combobox">
+          role="combobox"
+        >
           <div class="searchbox__container">
             <div class={searchBoxContainerClass} role="none">
               <template if:true={withoutSubmitButton}>
                 <lightning-icon
-                  class="slds-icon slds-input__icon slds-input__icon_left"
+                  class="slds-icon slds-input__icon slds-input__icon_left searchbox__search-icon"
                   icon-name="utility:search"
-                  size="x-small"></lightning-icon>
+                  size="x-small"
+                ></lightning-icon>
               </template>
               <textarea
                 id="combobox-with-button-id"
@@ -29,44 +30,52 @@
                 aria-autocomplete="list"
                 autocomplete="off"
                 role="textbox"
-                style="white-space: nowrap;">
-                </textarea>
+                style="white-space: nowrap"
+                value={value}
+              >
+              </textarea>
+
               <template if:false={isQueryEmpty}>
-                <div class="slds-grid slds-align_absolute-center slds-var-m-right_xxx-small searchbox__clear-button-container">
+                <div
+                  class="slds-grid slds-align_absolute-center slds-var-m-right_xxx-small searchbox__clear-button-container"
+                >
                   <lightning-button-icon
                     onclick={clearInput}
                     icon-name="utility:close"
-                    class="slds-grid searchbox__clear-button"
+                    class="slds-grid searchbox__clear-button-expandable"
                     variant="bare"
                     size="medium"
                     name={labels.clear}
                     title={labels.clear}
                     aria-label={labels.clear}
-                    alternative-text={labels.clear}></lightning-button-icon>
+                    alternative-text={labels.clear}
+                  ></lightning-button-icon>
                 </div>
               </template>
+
               <template if:false={withoutSubmitButton}>
                 <button
                   class="searchbox__submit-button slds-button slds-button_neutral"
                   title={labels.search}
-                  onclick={onSubmit}>
+                  onclick={onSubmit}
+                >
                   <lightning-icon
                     class="slds-current-color"
                     size="x-small"
                     icon-name="utility:search"
                     alternative-text={labels.search}
-                    title={labels.search}></lightning-icon>
+                    title={labels.search}
+                  ></lightning-icon>
                   <span class="slds-assistive-text">{labels.search}</span>
                 </button>
               </template>
             </div>
           </div>
-
           <c-quantic-search-box-suggestions-list
             suggestions={suggestions}
             onhighlightchange={handleHighlightChange}
-            onsuggestionselected={handleSuggestionSelection}>
-          </c-quantic-search-box-suggestions-list>
+            onsuggestionselected={handleSuggestionSelection}
+          ></c-quantic-search-box-suggestions-list>
         </div>
       </div>
     </div>
