+++ packages/quantic/force-app/main/default/lwc/quanticStandaloneSearchBox/quanticStandaloneSearchBox.js
@@ -1,4 +1,3 @@
-import clear from '@salesforce/label/c.quantic_Clear';
 import search from '@salesforce/label/c.quantic_Search';
 import {
   registerComponentForInit,
@@ -6,20 +5,13 @@ import {
   getHeadlessBindings,
   destroyEngine,
 } from 'c/quanticHeadlessLoader';
-import {STANDALONE_SEARCH_BOX_STORAGE_KEY, keys} from 'c/quanticUtils';
+import {STANDALONE_SEARCH_BOX_STORAGE_KEY} from 'c/quanticUtils';
 import {CurrentPageReference, NavigationMixin} from 'lightning/navigation';
 import {LightningElement, api, track, wire} from 'lwc';
 // @ts-ignore
-import defaultStandaloneSearchBox from './templates/defaultStandaloneSearchBox.html';
+import standaloneSearchBox from './templates/standaloneSearchBox.html';
 // @ts-ignore
 import errorTemplate from './templates/errorTemplate.html';
-// @ts-ignore
-import expandableStandaloneSearchBox from './templates/expandableStandaloneSearchBox.html';
-
-const CLASS_WITH_SUBMIT =
-  'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_right slds-input-has-fixed-addon';
-const CLASS_WITHOUT_SUBMIT =
-  'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_left-right';
 
 /** @typedef {import("coveo").SearchEngine} SearchEngine */
 /** @typedef {import("coveo").StandaloneSearchBoxState} StandaloneSearchBoxState */
@@ -39,7 +31,6 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
 ) {
   labels = {
     search,
-    clear,
   };
 
   /**
@@ -137,10 +128,11 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
 
   connectedCallback() {
     registerComponentForInit(this, this.standaloneEngineId);
-    this.addEventListener(
-      'suggestionlistrender',
-      this.handleSuggestionListEvent
-    );
+
+    this.addEventListener('inputValueChange', this.handleInputValueChange);
+    this.addEventListener('submitSearch', this.handleSubmit);
+    this.addEventListener('showSuggestions', this.handleShowSuggestions);
+    this.addEventListener('selectSuggestion', this.handleSelectSuggestion);
   }
 
   renderedCallback() {
@@ -186,10 +178,11 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
 
   disconnectedCallback() {
     this.unsubscribe?.();
-    this.removeEventListener(
-      'suggestionlistrender',
-      this.handleSuggestionListEvent
-    );
+
+    this.removeEventListener('inputValueChange', this.handleInputValueChange);
+    this.removeEventListener('submitSearch', this.handleSubmit);
+    this.removeEventListener('showSuggestions', this.handleShowSuggestions);
+    this.removeEventListener('selectSuggestion', this.handleSelectSuggestion);
   }
 
   updateStandaloneState() {
@@ -220,147 +213,40 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
     this.navigateToSearchPage();
   }
 
-  get searchBoxContainerClass() {
-    if (this.withoutSubmitButton) {
-      this.input?.setAttribute('aria-labelledby', 'fixed-text-label');
-      return CLASS_WITHOUT_SUBMIT;
-    }
-    this.input?.setAttribute(
-      'aria-labelledby',
-      'fixed-text-label fixed-text-addon-post'
-    );
-    return CLASS_WITH_SUBMIT;
-  }
-
-  get searchBoxInputClass() {
-    return `slds-input searchbox__input ${
-      this.withoutSubmitButton ? '' : 'searchbox__input-with-button'
-    }`;
-  }
-
-  showSuggestions() {
-    this.standaloneSearchBox?.showSuggestions();
-    this.combobox?.classList.add('slds-is-open');
-    this.combobox?.setAttribute('aria-expanded', 'true');
-  }
-
-  hideSuggestions() {
-    this.combobox?.classList.remove('slds-is-open');
-    this.combobox?.setAttribute('aria-expanded', 'false');
-    this.suggestionList?.resetSelection();
-  }
-
-  handleHighlightChange(event) {
-    this.input.value = event.detail?.rawValue;
-  }
-
-  handleEnter() {
-    const selectedSuggestion = this.suggestionList?.getCurrentSelectedValue();
-    if (this.suggestionsOpen && selectedSuggestion) {
-      this.standaloneSearchBox.selectSuggestion(selectedSuggestion.rawValue);
-    } else {
-      this.standaloneSearchBox.submit();
-    }
-    this.input.blur();
-  }
-
-  handleValueChange() {
-    if (this.standaloneSearchBox.state.value !== this.input.value) {
-      this.standaloneSearchBox.updateText(this.input.value);
-    }
-  }
-
-  onSubmit(event) {
-    event.stopPropagation();
+  // CustomEvent handlers
+  /**
+   * Updates the input value.
+   */
+  handleInputValueChange = (event) => {
+    const updatedValue = event.detail.newInputValue;
     if (this.standaloneSearchBox?.state?.value !== this.input.value) {
-      this.standaloneSearchBox.updateText(this.input.value);
+      this.standaloneSearchBox.updateText(updatedValue);
     }
+  };
+
+  /**
+   * Submits a search.
+   * @returns {void}
+   */
+  handleSubmit = () => {
     this.standaloneSearchBox.submit();
-    this.input.blur();
-  }
-
-  handleKeyValues() {
-    if (this.standaloneSearchBox?.state?.value !== this.input.value) {
-      this.suggestionList?.resetSelection();
-      this.standaloneSearchBox?.updateText(this.input.value);
-    }
-  }
+  };
 
   /**
-   * Prevent default behavior of enter key, on textArea, to prevent skipping a line.
-   * @param {KeyboardEvent} event
+   * Shows the suggestions.
+   * @returns {void}
    */
-  onKeydown(event) {
-    if (event.key === keys.ENTER) {
-      event.preventDefault();
-    }
-  }
+  handleShowSuggestions = () => {
+    this.standaloneSearchBox?.showSuggestions();
+  };
 
   /**
-   * @param {KeyboardEvent} event
+   * Handles the selection of a suggestion.
    */
-  onKeyup(event) {
-    switch (event.key) {
-      case keys.ENTER:
-        this.handleEnter();
-        break;
-      case keys.ARROWUP:
-        this.suggestionList.selectionUp();
-        break;
-      case keys.ARROWDOWN:
-        this.suggestionList.selectionDown();
-        break;
-      default:
-        this.handleKeyValues();
-    }
-  }
-
-  onFocus() {
-    this.showSuggestions();
-    this.adjustTextAreaHeight();
-  }
-
-  onBlur() {
-    this.hideSuggestions();
-    this.collapseTextArea();
-  }
-
-  onTextAreaInput() {
-    this.handleValueChange();
-    this.adjustTextAreaHeight();
-  }
-
-  adjustTextAreaHeight() {
-    if (!this.textarea) {
-      return;
-    }
-    this.input.value = this.input.value.replace(/\n/g, '');
-    this.input.style.height = '';
-    this.input.style.whiteSpace = 'pre-wrap';
-    this.input.style.height = this.input.scrollHeight + 'px';
-  }
-
-  collapseTextArea() {
-    if (!this.textarea) {
-      return;
-    }
-    this.input.style.height = '';
-    this.input.style.whiteSpace = 'nowrap';
-  }
-
-  clearInput() {
-    this.input.value = '';
-    this.standaloneSearchBox.updateText(this.input.value);
-    this.input.focus();
-    if (this.textarea) {
-      this.adjustTextAreaHeight();
-    }
-  }
-
-  handleSuggestionSelection(event) {
-    const textValue = event.detail;
-    this.standaloneSearchBox.selectSuggestion(textValue);
-  }
+  handleSelectSuggestion = (event) => {
+    const selectedSuggestion = event.detail.selectedSuggestion.rawValue;
+    this.standaloneSearchBox?.selectSuggestion(selectedSuggestion);
+  };
 
   resetStandaloneSearchboxState() {
     const engine = getHeadlessBindings(this.standaloneEngineId)?.engine;
@@ -387,49 +273,18 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
     );
   }
 
-  /**
-   * @returns {HTMLInputElement|HTMLTextAreaElement}
-   */
-  get input() {
-    return this.textarea
-      ? this.template.querySelector('textarea')
-      : this.template.querySelector('input');
-  }
+  // /**
+  //  * @returns {HTMLInputElement|HTMLTextAreaElement}
+  //  */
+  // get input() {
+  //   return this.textarea
+  //     ? this.template.querySelector('textarea')
+  //     : this.template.querySelector('input');
+  // }
 
-  /**
-   * @returns {HTMLElement}
-   */
-  get combobox() {
-    return this.template.querySelector('.slds-combobox');
-  }
-
-  /**
-   * @returns {quanticSearchBoxSuggestionsList}
-   */
-  get suggestionList() {
-    // @ts-ignore
-    return this.template.querySelector('c-quantic-search-box-suggestions-list');
-  }
-
-  /**
-   * @returns {boolean}
-   */
-  get isQueryEmpty() {
-    return !this.input?.value?.length;
-  }
-
-  /**
-   * @returns {boolean}
-   */
-  get suggestionsOpen() {
-    return this.combobox?.classList.contains('slds-is-open');
-  }
-
-  handleSuggestionListEvent = (event) => {
-    event.stopPropagation();
-    const id = event.detail;
-    this.input.setAttribute('aria-controls', id);
-  };
+  // get quanticSearchBoxInput() {
+  //   return this.template.querySelector('c-quantic-search-box-input');
+  // }
 
   /**
    * Sets the component in the initialization error state.
@@ -439,11 +294,6 @@ export default class QuanticStandaloneSearchBox extends NavigationMixin(
   }
 
   render() {
-    if (this.hasInitializationError) {
-      return errorTemplate;
-    }
-    return this?.textarea
-      ? expandableStandaloneSearchBox
-      : defaultStandaloneSearchBox;
+    return this.hasInitializationError ? errorTemplate : standaloneSearchBox;
   }
 }
