--- packages/quantic/force-app/main/default/lwc/quanticStandaloneSearchBox/templates/expandableStandaloneSearchBox.html
@@ -1,85 +1,0 @@
-<!-- sldsValidatorIgnore -->
-<template>
-  <template if:true={isStandalone}>
-    <c-quantic-search-interface engine-id={standaloneEngineId} search-hub={searchHub} pipeline={pipeline} skip-first-search disable-state-in-url>
-      <div class="slds-form-element z-index-high searchbox__container-wrapper">
-        <div class="slds-form-element__control searchbox_floating-container">
-          <div class="slds-combobox_container">
-            <div
-              class="slds-combobox slds-dropdown-trigger slds-dropdown-trigger_click"
-              aria-expanded="true"
-              aria-haspopup="listbox"
-              role="combobox">
-              <div class="searchbox__container">
-                <div class={searchBoxContainerClass} role="none">
-                  <template if:true={withoutSubmitButton}>
-                    <lightning-icon
-                      class="slds-icon slds-input__icon slds-input__icon_left searchbox__search-icon"
-                      icon-name="utility:search"
-                      size="x-small"></lightning-icon>
-                  </template>
-                  <textarea
-                    id="combobox-with-button-id"
-                    class={searchBoxInputClass}
-                    placeholder={placeholder}
-                    type="search"
-                    onkeydown={onKeydown}
-                    onkeyup={onKeyup}
-                    onfocus={onFocus}
-                    onblur={onBlur}
-                    oninput={onTextAreaInput}
-                    aria-autocomplete="list"
-                    autocomplete="off"
-                    role="textbox"
-                    style="white-space: nowrap;">
-                </textarea>
-                  <template if:false={isQueryEmpty}>
-                    <div class="slds-grid slds-align_absolute-center slds-var-m-right_xxx-small searchbox__clear-button-container">
-                      <lightning-button-icon
-                        onclick={clearInput}
-                        icon-name="utility:close"
-                        variant="bare"
-                        class="slds-grid searchbox__clear-button"
-                        size="medium"
-                        name={labels.clear}
-                        title={labels.clear}
-                        aria-label={labels.clear}
-                        alternative-text={labels.clear}></lightning-button-icon>
-                    </div>
-                  </template>
-                </div>
-                <template if:false={withoutSubmitButton}>
-                  <button 
-                    class="searchbox__submit-button slds-button slds-button_neutral" 
-                    title={labels.search} 
-                    onclick={onSubmit}>
-                    <lightning-icon 
-                      class="slds-current-color" 
-                      size="x-small" 
-                      icon-name="utility:search" 
-                      alternative-text={labels.search} 
-                      title={labels.search}></lightning-icon>
-                    <span class="slds-assistive-text">{labels.search}</span>
-                  </button>
-                </template>
-              </div>
-
-              <c-quantic-search-box-suggestions-list
-                suggestions={suggestions}
-                onhighlightchange={handleHighlightChange}
-                onsuggestionselected={handleSuggestionSelection}></c-quantic-search-box-suggestions-list>
-            </div>
-          </div>
-        </div>
-      </div>
-    </c-quantic-search-interface>
-  </template>
-  <template if:false={isStandalone}>
-    <c-quantic-search-box
-      engine-id={engineId}
-      number-of-suggestions={numberOfSuggestions}
-      placeholder={placeholder}
-      without-submit-button={withoutSubmitButton}
-      textarea={textarea}></c-quantic-search-box>
-  </template>
-</template>
