+++ packages/quantic/force-app/main/default/lwc/quanticStandaloneSearchBox/templates/standaloneSearchBox.html
@@ -1,0 +1,28 @@
+<template>
+  <template if:true={isStandalone}>
+    <c-quantic-search-interface
+      engine-id={standaloneEngineId}
+      search-hub={searchHub}
+      pipeline={pipeline}
+      skip-first-search
+      disable-state-in-url
+    >
+      <c-quantic-search-box-input
+        textarea={textarea}
+        input-value={inputValue}
+        without-submit-button={withoutSubmitButton}
+        placeholder={placeholder}
+        suggestions={suggestions}
+      ></c-quantic-search-box-input>
+    </c-quantic-search-interface>
+  </template>
+  <template if:false={isStandalone}>
+    <c-quantic-search-box
+      engine-id={engineId}
+      placeholder={placeholder}
+      without-submit-button={withoutSubmitButton}
+      number-of-suggestions={numberOfSuggestions}
+      textarea={textarea}
+    ></c-quantic-search-box>
+  </template>
+</template>
